DDS-341 simplify rtps types (#162)
* replace GuidPrefix by typedef

* simplify EntityKey

* use OctetArray3

* Simplify EntityKind

* Make SequenceNumber as in IDL

* refactor LocatorKind

* refactor LocatorPort

* fix clippy

* refactor ProtocolVersion

* remove enum DurabilityKind

* remove ExpectsInlineQos

* remove redundant WriteBytes impls

* simplify Count

* refactor ParameterId

* simplify FragmentNumber_t

* clean types

* fix clippy

+++ dds/src/implementation/dds/message_receiver.rs
@@ -45,7 +45,7 @@ impl<'a> Iterator for MessageReceiver<'a> {
                     if !m.invalidate_flag() {
                         self.have_timestamp = true;
                         self.timestamp =
-                            Time::new(m.timestamp().seconds(), m.timestamp().fraction());
+                            Time::new(m.timestamp().seconds() as i32, m.timestamp().fraction());
                     } else {
                         self.have_timestamp = false;
                         self.timestamp = TIME_INVALID;
